Reduce unit test workload.
Don't run tests on both service connection types unless the test
actually cares about the service connection type.
