+++ tests/envs/llvm/service_connection_test.py
@@ -1,0 +1,71 @@
+# Copyright (c) Facebook, Inc. and its affiliates.
+#
+# This source code is licensed under the MIT license found in the
+# LICENSE file in the root directory of this source tree.
+"""Integrations tests for the LLVM CompilerGym environments."""
+from pathlib import Path
+
+import gym
+import numpy as np
+import pytest
+
+import compiler_gym  # Register environments.
+from compiler_gym.envs import CompilerEnv
+from compiler_gym.envs.llvm.llvm_env import LlvmEnv
+from compiler_gym.service.connection import CompilerGymServiceConnection
+from compiler_gym.util.runfiles_path import runfiles_path
+from tests.test_main import main
+
+pytest_plugins = ["tests.envs.llvm.fixtures"]
+
+SERVICE_BIN = Path(runfiles_path("CompilerGym/compiler_gym/envs/llvm/service/service"))
+
+
+@pytest.fixture(scope="function", params=["local", "service"])
+def env(request) -> CompilerEnv:
+    # Redefine fixture to test both gym.make(...) and unmanaged service
+    # connections.
+    if request.param == "local":
+        env = gym.make("llvm-v0")
+        env.require_dataset("cBench-v0")
+        try:
+            yield env
+        finally:
+            env.close()
+    else:
+        service = CompilerGymServiceConnection(SERVICE_BIN)
+        env = LlvmEnv(service=service.connection.url, benchmark="foo")
+        env.require_dataset("cBench-v0")
+        try:
+            yield env
+        finally:
+            env.close()
+            service.close()
+
+
+def test_service_env_dies_reset(env: CompilerEnv):
+    env.observation_space = "Autophase"
+    env.reward_space = "IrInstructionCount"
+    env.reset("cBench-v0/crc32")
+
+    # Kill the service.
+    env.service.close()
+
+    # Check that the environment doesn't fall over.
+    observation, reward, done, _ = env.step(0)
+    assert done
+
+    # Check that default values are returned.
+    np.testing.assert_array_equal(observation, np.zeros(56))
+    assert reward == 0
+
+    # Reset the environment and check that it works.
+    env.reset(benchmark="cBench-v0/crc32")
+    observation, reward, done, _ = env.step(0)
+    assert not done
+    assert observation is not None
+    assert reward is not None
+
+
+if __name__ == "__main__":
+    main()
